Fix System.IO.FileSystem product source and tests for Linux
Prior to this set of changes, 110 of the FileSystem tests were failing on Linux.  After this change, all allowed tests pass.

Primary product fixes:
- Using rmdir/unlink for directory/file deletion rather than using remove
- Disallowing FileStreams to be opened for directories
- Supporting copying files when only target directory (not target name) is specified
- Overhauling directory enumeration to ensure to match Win32 behaviors related to search pattern normalization, input validation, supporting re-iteration through an enumerable, outputting partial paths based on user-supplied path, using a SafeHandle to ensure resources are cleaned up
- Fixing FileSystemObject.LastWriteTime to correctly return the LastWriteTime rather than LastAccessTime
- Fixing FileAttributes.ReadOnly handling due to inappropriate caching
- Validating inputs to FileSystemObject.Attributes, and make sure that changes to attributes are visible afterwards
- Adding support for getting CreationTime if it's supported by the OS
- Marking links with FileAttributes.ReparsePoint
- Working around current runtime bug related to exception handling and finally blocks getting executed twice

Primary test fixes:
- Suppressing tests as platform-specific that dependend on Win32-isms like drive letters, alternate data streams, etc.
- Changing a bunch of '\' usage to instead use Path.DirectorySeparatorChar or Path.Combine
- Adapting to behavioral differences between Windows and Unix, e.g. whether a directory can be deleted if it's readonly, invalid path characters, max path lengths, file sharing differences, etc.

These changes should all accrue to OSX as well, but we'll need to separately do an exercise to clean this library and tests up there for anything OSX-specific.

+++ src/System.IO.FileSystem/tests/Directory/Move_str_str.cs
@@ -197,33 +197,36 @@ public class Directory_Move_str_str
             Directory.Delete(tempDirName);
 
 #if !TEST_WINRT
-            // [] Move to different drive will throw AccessException
-            //-----------------------------------------------------------------
-            strLoc = "Loc_00025";
-            string fullDirName = Path.GetFullPath(dirName);
-            Directory.CreateDirectory(dirName);
-            iCountTestcases++;
-            try
+            if (Interop.IsWindows) // moving between drive labels
             {
-                if (fullDirName.Substring(0, 3) == @"d:\" || fullDirName.Substring(0, 3) == @"D:\")
-                    Directory.Move(fullDirName, "C:\\TempDirectory");
-                else
-                    Directory.Move(fullDirName, "D:\\TempDirectory");
-                Console.WriteLine("Root directory..." + fullDirName.Substring(0, 3));
-                iCountErrors++;
-                printerr("Error_00026! Expected exception not thrown");
+                // [] Move to different drive will throw AccessException
+                //-----------------------------------------------------------------
+                strLoc = "Loc_00025";
+                string fullDirName = Path.GetFullPath(dirName);
+                Directory.CreateDirectory(dirName);
+                iCountTestcases++;
+                try
+                {
+                    if (fullDirName.Substring(0, 3) == @"d:\" || fullDirName.Substring(0, 3) == @"D:\")
+                        Directory.Move(fullDirName, "C:\\TempDirectory");
+                    else
+                        Directory.Move(fullDirName, "D:\\TempDirectory");
+                    Console.WriteLine("Root directory..." + fullDirName.Substring(0, 3));
+                    iCountErrors++;
+                    printerr("Error_00026! Expected exception not thrown");
+                }
+                catch (IOException)
+                {
+                }
+                catch (Exception exc)
+                {
+                    iCountErrors++;
+                    printerr("Error_00000! Incorrect exception thrown, exc==" + exc.ToString());
+                }
+                if (Directory.Exists(fullDirName))
+                    Directory.Delete(fullDirName, true);
+                //-----------------------------------------------------------------
             }
-            catch (IOException)
-            {
-            }
-            catch (Exception exc)
-            {
-                iCountErrors++;
-                printerr("Error_00000! Incorrect exception thrown, exc==" + exc.ToString());
-            }
-            if (Directory.Exists(fullDirName))
-                Directory.Delete(fullDirName, true);
-            //-----------------------------------------------------------------
 #endif
 
             // [] Moving Directory with subdirectories
@@ -306,7 +309,7 @@ public class Directory_Move_str_str
             iCountTestcases++;
             try
             {
-                Directory.Move(TestInfo.CurrentDirectory, "<MyDirectory>");
+                Directory.Move(TestInfo.CurrentDirectory, "<MyDirectory\0");
                 iCountErrors++;
                 printerr("Error_00041! Expected exception not thrown");
             }
@@ -324,9 +327,7 @@ public class Directory_Move_str_str
             //-----------------------------------------------------------------
             strLoc = "Loc_00044";
 
-            String str = "";
-            for (int i = 0; i < 250; i++)
-                str += "a";
+            String str = new string('a', IOInputs.MaxPath);
             iCountTestcases++;
             try
             {
@@ -349,23 +350,26 @@ public class Directory_Move_str_str
             //-----------------------------------------------------------------
             strLoc = "Loc_00048";
 
-            iCountTestcases++;
-            Directory.CreateDirectory(dirName);
-            try
+            if (Interop.IsWindows) // drive labels
             {
-                Directory.Move(dirName, "X:\\Temp");
-                iCountErrors++;
-                printerr("Error_00049! Expected exception not thrown");
+                iCountTestcases++;
+                Directory.CreateDirectory(dirName);
+                try
+                {
+                    Directory.Move(dirName, "X:\\Temp");
+                    iCountErrors++;
+                    printerr("Error_00049! Expected exception not thrown");
+                }
+                catch (IOException)
+                {
+                }
+                catch (Exception exc)
+                {
+                    iCountErrors++;
+                    printerr("Error_00051! Incorrect exception thrown, exc==" + exc.ToString());
+                }
+                Directory.Delete(dirName);
             }
-            catch (IOException)
-            {
-            }
-            catch (Exception exc)
-            {
-                iCountErrors++;
-                printerr("Error_00051! Incorrect exception thrown, exc==" + exc.ToString());
-            }
-            Directory.Delete(dirName);
             //-----------------------------------------------------------------
 
             // [] Use directory names with spaces
